% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender_demographic.R
\name{get_gender_demographic}
\alias{get_gender_demographic}
\title{Returns the gender distribution of visitors by month for a given website}
\usage{
get_gender_demographic(
  api_key = similarweb_api_key,
  site,
  start,
  end,
  country = "us",
  md = "False"
)
}
\arguments{
\item{api_key}{API key is required.
Please set API key equal to similarweb_api_key before calling function.
If you do not have an API key, you can get an API key at https://account.similarweb.com/#/api-management2
Note: You must have a subscription to SimilarWeb's services.}

\item{site}{Website name is required. Do not include 'www' or 'https' before website. Ex. 'google.com'}

\item{start}{Start date required in YYYY-MM format. Ex. '2020-10'}

\item{end}{End date required in YYYY-MM format. Ex. '2020-12'}

\item{country}{Country set to 'US' as default but can be set to another 2-letter ISO country code depending on your level of access within SimilarWeb.}

\item{md}{Default is set to 'False' but can be set to 'True' for main domain only}
}
\value{
tibble with gender and value columns
}
\description{
The `get_gender_demographic()` function returns the male/female distribution of visitors to a given website based on the start and end date you provide.
}
\examples{
##gender_demographic <- get_gender_demographic(site = "bbc.com", start = "2020-10", end = "2020-12")
}
