% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similar_sites.R
\name{get_similar_sites}
\alias{get_similar_sites}
\title{Returns the top sites that are most similar to the given website}
\usage{
get_similar_sites(api_key = similarweb_api_key, site, md = "False")
}
\arguments{
\item{api_key}{API key is required.
Please set API key equal to similarweb_api_key before calling function.
If you do not have an API key, you can get an API key at https://account.similarweb.com/#/api-management2
Note: You must have a subscription to SimilarWeb's services.}

\item{site}{Website name is required. Do not include 'www' or 'https' before website. Ex. 'google.com'}

\item{md}{Default is set to 'False' but can be set to 'True' for main domain only}
}
\value{
tibble with site url and similarity score columns
}
\description{
The `get_similar_sites()` function returns the top sites that are most similar to the given website within the past 28 days.
}
\examples{
##top_similar_sites <- get_similar_sites(site = "bbc.com")
}
